t13: largeintbuffer in datamanager
zeiten sind aber durch auslagerung verfälscht, lief hart an der grenze
                        PC: Sven-LAPTOP
    Beginne Sortierung von: 1.000.000.000 Integers (3814MB)
Initial Integers pro Block: 640.679.936
   Integers pro Merge-Read: 33.554.432
  Integers pro Merge-Write: 33.554.432
       Max Arbeitsspeicher: 2.444MB
 Max GrÃ¶ÃŸe eines Readcalls: 128MB

readcall: 48588 ms
Start Quicksort
Stop Quicksort
quicksort: 115722 ms
writecall: 46666 ms
readcall: 42363 ms
Start Quicksort
Stop Quicksort
quicksort: 62415 ms
writecall: 50029 ms
readcall: 16 ms
06:22 - Diff 382.9s: QuickSort abgschlossen. Beginne ersten Merge-Schrit mit RunlÃ¤nge = 640.679.936
09:25 - Diff 182.0s: Sortieren abgeschlossen! Die Ausgabedatei enthÃ¤lt 1.000.000.000 Integers.

 ----- hier nochmal der normale buffer ----------

                        PC: Sven-LAPTOP
    Beginne Sortierung von: 1.000.000.000 Integers (3814MB)
Initial Integers pro Block: 500.000.001
   Integers pro Merge-Read: 33.554.432
  Integers pro Merge-Write: 33.554.432
       Max Arbeitsspeicher: 2.047MB
 Max GrÃ¶ÃŸe eines Readcalls: 128MB

readcall: 37990 ms
Start Quicksort
Stop Quicksort
quicksort: 46542 ms
writecall: 34120 ms
readcall: 34935 ms
Start Quicksort
Stop Quicksort
quicksort: 44739 ms
writecall: 33635 ms
readcall: 0 ms
03:53 - Diff 233.5s: QuickSort abgschlossen. Beginne ersten Merge-Schrit mit RunlÃ¤nge = 500.000.001
06:23 - Diff 150.2s: Sortieren abgeschlossen! Die Ausgabedatei enthÃ¤lt 1.000.000.000 Integers.

+++ trunk/ad-praktikum-t13-t14/src/aufgabe2/data/io/Reader.java
@@ -22,7 +22,7 @@ public class Reader {
 
     //Konstanten
     private final long fileChanSize; //Die Größe der Datei   
-    private final int bytesPerRead; //Die anzahl der Elemente, die pro readToByteBuffer in den Buffer gelesen werden sollen
+    private final long bytesPerRead; //Die anzahl der Elemente, die pro readToByteBuffer in den Buffer gelesen werden sollen
     //private final int internalMaxReadSize; //Noch nicht benutzt: Die maximale Anzahl der auf einmal eingelesenen Bytes für fileChan.read. Ist bBufferSize größer als dieser Wert, so wird ggf. mehrmals hintereinander gelesen, bis readToByteBuffer den kompletten Wert zurückliefert. 
         
     /**
@@ -31,7 +31,7 @@ public class Reader {
      * @param bufferSize
      * @return
      */
-    public static Reader create(String fileName, int bufferSize){
+    public static Reader create(String fileName, long bufferSize){
         try {
 			return new Reader(fileName, bufferSize);
 		} catch (IOException e) {
@@ -45,7 +45,7 @@ public class Reader {
      * @param bBufferSize
      * @throws IOException 
      */
-    private Reader(String filePath, int bBufferSize) throws IOException{
+    private Reader(String filePath, long bBufferSize) throws IOException{
         if (bBufferSize % Constants.INTSIZE != 0)
         	throw new IllegalArgumentException("BuferSize muss ein Vielfaches von INTSIZE sein.");
             
@@ -121,8 +121,8 @@ public class Reader {
      * @throws IOException 
      */
     public void readToByteBuffer(ByteBuffer target) throws IOException {
-        int bytesToRead = (int)Math.min(bytesPerRead, fileChanSize- currentCursorPosition); //Insgesammt zu lesende Menge an Bytes
-        int currentReadSize = 0;
+        long bytesToRead = Math.min(bytesPerRead, fileChanSize- currentCursorPosition); //Insgesammt zu lesende Menge an Bytes
+        long currentReadSize = 0;
         if (target.capacity() < bytesToRead){
 //            throw new IllegalArgumentException("Übergebener Buffer zu klein");
             currentReadSize = Math.min(target.capacity(), bytesToRead);
@@ -130,7 +130,7 @@ public class Reader {
             target.clear();
 
             //Limit neu setzen
-            target.limit(target.position() + currentReadSize);
+            target.limit((int)(target.position() + currentReadSize));
 
             //Lesen
             fileChan.read(target, currentCursorPosition);
@@ -146,7 +146,7 @@ public class Reader {
                 currentReadSize = Math.min(Constants.MAXBYTESPERREADCALL, bytesToRead);
 
                 //Limit neu setzen
-                target.limit(target.position() + currentReadSize);
+                target.limit((int)(target.position() + currentReadSize));
 
                 //Lesen
                 fileChan.read(target, currentCursorPosition);
